divide OpenGL into
